Background
Added a border so that i can get in my head some of the changes I'm making and how I'm going to make them. Should make compartmentalizing things a little easier. Figured out how to stretch the Toast image all the way through the background but can't tell if I like it or not. Might try to go for a star design instead. Made some other small changes.